Extract npm simulationsAndTryDifferentNumbersOfParallelWorkers script
