+++ src/scripts/simulationsAndTryDifferentNumbersOfParallelWorkers.ts
@@ -1,0 +1,3 @@
+import { runSimulationsAndTryDifferentNumbersOfParallelWorkers } from '../runSimulationsAndTryDifferentNumbersOfParallelWorkers';
+
+runSimulationsAndTryDifferentNumbersOfParallelWorkers();
